// max-mini-app/src/api/models/QueueMember.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MaxQueue Orchestrator API
 * API, which is the entry point to the MaxQueue project
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface QueueMember
 */
export interface QueueMember {
    /**
     * 
     * @type {number}
     * @memberof QueueMember
     */
    maxId?: number;
    /**
     * 
     * @type {string}
     * @memberof QueueMember
     */
    queueEntryId?: string;
    /**
     * 
     * @type {string}
     * @memberof QueueMember
     */
    username?: string;
}

/**
 * Check if a given object implements the QueueMember interface.
 */
export function instanceOfQueueMember(value: object): value is QueueMember {
    return true;
}

export function QueueMemberFromJSON(json: any): QueueMember {
    return QueueMemberFromJSONTyped(json, false);
}

export function QueueMemberFromJSONTyped(json: any, ignoreDiscriminator: boolean): QueueMember {
    if (json == null) {
        return json;
    }
    return {
        
        'maxId': json['maxId'] == null ? undefined : json['maxId'],
        'queueEntryId': json['queueEntryId'] == null ? undefined : json['queueEntryId'],
        'username': json['username'] == null ? undefined : json['username'],
    };
}

export function QueueMemberToJSON(json: any): QueueMember {
    return QueueMemberToJSONTyped(json, false);
}

export function QueueMemberToJSONTyped(value?: QueueMember | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'maxId': value['maxId'],
        'queueEntryId': value['queueEntryId'],
        'username': value['username'],
    };
}